use dedicated file
